backend/requirements.txt
fastapi==0.115.6
pydantic-settings==2.7.1
pydantic==2.10.5
asyncpg==0.30.0
httpx==0.27.2
python-telegram-bot==21.10
python-dotenv==1.0.1
psycopg2-binary==2.9.10
uvicorn[standard]==0.34.0

